feat(a11y): moves to a component in baseui
