+++ src/a11y/index.js
@@ -1,0 +1,9 @@
+/*
+Copyright (c) 2018-2019 Uber Technologies, Inc.
+
+This source code is licensed under the MIT license found in the
+LICENSE file in the root directory of this source tree.
+*/
+// @flow
+
+export {default as Unstable_A11y} from './a11y.js';
